Added support for Access 2007 and earlier
The Declare statements use the PtrSafe keyword which doesn't exist
in Access 2007. So, they need to be wrapped in a if-else that checks
the version of VBE.
See: https://msdn.microsoft.com/en-us/library/office/gg278832.aspx

The enumeration value  is a new member of .
Data macros are a new feature of Access 2010.
See: https://support.office.com/en-us/article/Create-a-data-macro-b1b94bca-4f17-47ad-a66d-f296ef834200