Code cleanup
· Correct scope of a procedure not explicitly declared
· Correct data access object declared without library
Specify which library to use, DAO library or ADODB library.
· Correct function declared without data type
Convert funtions on sub or specify correct type.
· Correct constant/variable declared without a data type
The default Variant data type is used. This data type is sloppy and slow.
· Correct variable 'Dim ... As New ...' declared
If you declare a variable this way you cannot check if this variable is Nothing.
· Correct '' instead of vbNullString used
You want to check if a string is empty. Better use the constant vbNullString.
This makes your code more meaningful.
· Correct Parameter not explicitly declared ByVal / ByRef.
· Change 'While-Wend' loop to 'While Do-Loop' loop
Using While-Wend loop is considered outdated.
· Correct funtions return string, 'fun' is used instead of 'fun$'
The function without the dollar sign returns as a result an expression of data
type Variant. The version with a dollar sign at the end returns as a result an
expression of data type string.
· Correct multiple commands divided by a colon.

by [TM VBA-Inspector] (http://www.team-moeller.de/en/)