Changed ExportTableDef to export the table schema in XML
The tableDef file are now exported as a table schema in XML as opposed to text with SQL. You no longer have to add ADO library references and remove default libraries due to conflicts. Exporting should work.

Importing might need some work. I see a potential problem if tables are imported out of order. Meaning it has a relation with a table that has not been created yet. A list may need to be created on export so we can import tables in the correct order.