Appended "VCS_" to all but the main Public  functions.
"VCS_" was added to remove naming collisions with existing functions. And easy target is "FileExists" as there are likely thousands of them on the web. This is even an issue with more obscure function names. I already had an "ExportTableDef" function beleave it or not (I replaced the old one with mine so beleave it). This is because well named descriptive functions and variables are self documenting code and good practice. All of this could be avoided if all the code was in a single module. It is easier to maintain as seperate mods but this tool is really guest code in another project and should be as low key as possible.

The main functions I mentioned are:  
- ExportAllSource
- ImportAllSource
- ImportProject
- make  

"VCS_" was appended to:
- ExportDataMacros
- ImportDataMacros
- ProjectPath
- MkDirIfNotExist
- DelIfExist
- ClearTextFilesFromDir
- FileExists
- ConvertUcs2Utf8
- ConvertUtf8Ucs2
- UsingUcs2
- TempFile
- ExportObject
- ImportObject
- SanitizeTextFiles
- ImportReferences
- ExportReferences
- ExportRelation
- ImportRelation
- ExportPrintVars
- ImportPrintVars
- Sb_Init
- Sb_Clear   (this isn't used and can go as well)
- Sb_Append
- Sb_Get
- PadRight
- ExportLinkedTable
- ExportTableDef
- ExportTableData
- ImportLinkedTable
- ImportTableDef
- ImportTableData