Create installer functions
Added functionality so that the user is prompted to install or upgrade the add-in when opening the add-in file. This creates a very user-friendly process for new users to get started with this tool. It also provides a simple upgrade process where they can simply download the file from GitHub and run it to update to the latest version.