New function to test UCS2 requirement
When importing into a blank database, we don't have any objects to export that we can test, so I took the simpler approach of creating a tiny query and exporting that as text to test USC2. In the case of an ADP project, I export a form since we don't have the lighter query option.
The result is cached so in a normal import scenario this would only be run once.