Import internal import/export specs
When you link a table it creates some records in two system tables to define the specification. This spec is not available through the UI, but is required to recreate some linked tables, such as a linked CSV file. This is now supported, and allows the export/import of linked text files that use an embedded specification.