Add DevMode class
This class will be used to manage the saved printer properties for forms and reports. The goal here is to be able to save and apply general printer settings like paper orientation and margins without needing to save the printer name if a specific printer was not specified for the report.