Only update JSON files when content changes
Rather than always writing the JSON files each time the export is run, it will now read the existing file and compare the `Items` collection to see if the actual content changed. This avoids constantly updating the json files when it is only the header that changed. (Most often because VCS was upgraded to a new build version.) This should help reduce the number of changed files in the development cycle.