Set encryption to NONE by default
At the suggestion of a user in #54 I am changing the default encryption to None. For most users working with internal or private repositories this should be fine. The risk is that someone might upload the source of their project to the Internet and not realize that they were sharing potentially sensitive details like their AD domain and user. But I think it is a reasonable expectation that the developer is responsible to understand those risks and review any files they are uploading to a public repository. Closes #54