Use relative paths for library references
Referenced Microsoft Access database libraries may be located in a relative path. This change allows us to build from source using relative paths. In my case the library databases are located (and deployed) in the same folder as the main database, so this resolves a build error when linking the VBE references.