Add option to extract theme files
The only time you really need to extract the *.thmx file is if you are customizing a theme and would like to track the customizations in version control. (You may want to consider tracking the theme in a different project, if it is shared across different databases.)
I have added a new option so that by default the theme files are not extracted. This should improve the export/import performance in almost every use case.