Always apply print settings when importing reports
Some print settings, such as page orientation are saved in the Printer object of the report. Since this is stripped out during the sanitization process, and the values stored in a JSON file, we need to always restore the print settings, even if the default printer is being used for the report. See #70