Progress on DevMode print settings
I am now able to build a PrtDevNames block from a printer object that is identical to the one that is saved with a report. We should be pretty close with PrtDevMode as well. Still a work in progress, but getting closer to writing these blocks back to the sanitized export files before import. See #70