Fix name brackets
A couple more places to fix for names to be properly handled. Thanks, @effjot for pointing out the one I missed, and @Indigo744 for highlighting the issue with the pseudo index. I believe both of these should be resolved now, but let me know if you encounter anything else!. See #83