Implement performance logging on Export
Added a new class to manage performance logging during export/import operations. This will be very helpful on large or complex databases where we want to see exactly where performance bottlenecks are occurring. A performance report is written to the end of the log file during the export or build operation.