Rewrote Sanitize function for major performance gains
I rewrote the Sanitize function from scratch, replacing the original complex RegEx expressions with simple string comparisons. In my opinion, this greatly increases the readability of the code when you don't have to try to understand RegEx expressions. It also reduces the overhead of creating and executing RegEx when string comparisons are much faster in this application. Performance testing before and after the rewrite was showing the new version running 5 to 10 times faster!