Add MidStr and RightStr functions to clsConcat
Added more functionality to the concatenation class so that we can efficiently pull parts of a string if we don't need the whole thing. (With very large strings, this is more efficient than using the VBA functions after `GetStr` because you don't have to reallocate memory for the entire string before grabbing a chunk of it.) An example is where the console display only loads the last 20,000 characters of the log.