Fix Fast Save for linked table definitions
These files are stored in JSON format, and were not being updated if the contents had not changed. This meant that the files were being rewritten each time instead of taking advantage of the Fast Save option. Now if Fast Save is enabled, the file is always written (if changed). If Fast Save is off, the file is not rewritten if the contents are the same.