Only save non-default printer name
The printer name is used in both DevMode and DevNames. When an Access report uses the default printer, the DevMode\Device name is blank. However, when building a DevMode structure from scratch using the default printer, there *will* be a device name returned. In this case, we should clear the device name before creating the PrtDevMode block for import. See #70