Fix Deploy function to properly export source
In the past, the deploy function would cause the working add-in database to become corrupted, require a rebuild from source files. I believe this was due to exporting modules while they were currently running code. This update finally fixes that by running the export from the installed add-in, rather than the current database. This involved a couple interesting techniques to load the add-in at the application level, and then to use an API Timer callback to launch the export *after* the code execution has finished in the current DB. It appears to work beautifully with no more crashes. :thumbsup: