Fix add-in source export
Setting the VBE CurrentProject doesn't always determine the context of the `Application.Run` function, so I was encountering issues where the export functions were being run in the current rpoject instead of the add-in. (Especially when developing the addin, since both VBProjects have the same name.) This change ensures that the export is running from the add-in project, not the current DB, and thereby avoiding the file/memory corruption issues.