Remove version numbers from GUIDs
Thanks to a great tip from @Indigo744 in #97, we don't need to include the VBE reference version numbers in version control, or specify them when adding the reference. If we simply use 0.0 as the GUID version, it will resolve to the latest installed version automagically. :smile: Closes #97. See also #96 for additional context.