Initial implementation of component index
The new `clsVCSIndex` class manages the saved state of the database components, which is an essential building block for the proposed improvements to fast save and git integration. I have implemented this on a few objects for export and import and I am pretty happy with the way it is working.