Implement Error Levels
This will allow us to more globally handle errors during an export or build operation. The Log class will maintain the overall error flag which represents the highest error encountered across the application. This can be checked during export and build operations to cancel the process if a critical error is encountered. #110