Reorganize functions into modules
Thanks to a good suggestion from @mmzmusicnotes, I reorganized many of the functions from modFunctions into new/other modules with more logical purposes. This should make it easier to find and organize functions going forward. **Note that this commit does not make any code changes** (other than a few scope changes), but strictly moves the existing code around. I did not make this change lightly, knowing that it will affect the history in GitHub, but I feel that it will be beneficial in the long term to have some better organization as the project continues to grow and improve over time.