Refine modification dates logic
It looks like the hash is going to be far more reliable for detecting changes in modules than trying to go by the `.DateModified` property of the `AccessObject`. It is also much faster than exporting every module each time the `.DateModified` property changes. However, this could be impacted by other module properties such as changing the class instancing. (Perhaps add that to the hash?)