Include code to add trusted location
In some environments, security settings may prevent the add-in from running in Microsoft Access. The `VerifyTrustedLocation` function will programatically add the Add-Ins folder as a trusted location through the registry, allowing the add-in to function normally. Since this seems to be an unusual case, I am not at this point running this as a part of the regular install routine since it could also impact the security of other add-ins that might exist in this folder. Closes #120