Don't add a text property with a null value
A database may contain a `DefaultBackupLocation` property with a null value if the database was backed up to a file in the same folder. Since we cannot create a text property with a null value, we will skip the import of this property. Fixes #126

(cherry picked from commit d030786ca57cd6e5bfb78d807c1518ab9c7ee065)