Integrate installer form, split installation functions to own module
Initial first pass in linking in the installer form to the installation functions. I also decided to split out the installer logic into `modInstall` so that `modAddin` can stay focused more narrowly on the actual operation of the add-in. More work and refactoring to do, but this gives us a baseline before I do much refactoring on the logic side. See #120