Implement index-based export for changed files on forms and reports
Now with FastSave enabled, forms and reports are verified with the index from the last export, and only exported if the modified date of the object is more recent than the last export, or if the VBA hash changes. This gets FastSave working correctly for forms and reports, and gives a good implementation example for other objects going forward.