Add hourglass on install
When clicking the install button, there may be a pause for a second or two before the install completes. Added an hourglass icon to indicate the busy state.

Also updated the MsgBox2 function to automatically turn off any hourglass before displaying a message, and restore the previous state after the message. (Simplifies the extra coding otherwise needed to check the hourglass state before every message displayed to the user.)