Don't create print settings files when using defaults
If the print settings of a form or report exactly matches the settings in the current default printer, there is no need to export the .json file with print settings. This should help eliminate a large number of unneeded .json files that would otherwise be tracked in version control. A cached hash of the current default printer settings is stored in VCSIndex at run time, but not saved to the index. Relates to #115