Improve UI experience after options change
If options have changed, then a full export will be required. This update sets the full export box as checked and disables it when a full export will take place anyway. This should be less confusing to the user who might not otherwise understand why a full export took place when the box was not checked.