Add support for alternative path separators
While `\` is commonly used as the path separator for English versions of Windows, this is not always the case with other languages. In this update I have tried to go through the code and replace all the hard-coded path separators with the one returned by `FSO.BuildPath`. This should allow the add-in to work in non-English environments where a different path separator is being used. May be a factor in issue #180