Use system code page for ANSI conversion
Due to problems with the `_autodetect_all` encoding parameter when processing source code files, we are now using a simple `GetACP` API call to retrieve the active code page from Windows, and mapping this to the appropriate encoding. (With support to easily add other encoding mappings in the future.) This should solve the issue described in #186 where the automatic detection was guessing the language incorrectly.