Fix read of zero line code modules
Switching this back to a number after an export failed when `.CountOfLines + .CountOfDeclarationLines = 0`. Yes, I could add `+1` but that is getting really long and confusing. I really don't think anyone is realistically going to put one billion lines in a single code module (even if you could exceed the limit of 65,534 lines). Let's keep it simple. :smile:  Relates to #191