Use case-insensitive comparison for table list
This add-in allows you to save data from a list of named tables. However, the names of the tables in other databases may use different casing. This change sets the table name comparison to case-insensitive so the name will still match even if the case is different. (This shouldn't create any side-effects since you can't have two tables with the same name and different casing in the same database.) Fixes #207