Discard RubberDuck Identifier in HelpFile field
RubberDuck evidently stores a project identifier in the HelpFile field of the VBE Project. This number seems to change randomly causing unwanted noise in version control, and really isn't a help file at all. This adds functionality to discard the ID by default, with an option to preserve it if someone really wants to. Closes #197