Add bootstrap feature to run VBA code before beginning build
In some cases a developer may want to run some custom code at the very beginning of the build process, before loading in all the references and objects. For example, in my environment we use Access database libraries in the same folder as the primary database. The libraries are not included in the VCS project, so when building from source, the files need to be copied to the local folder before the build process attempts to add them as references. This functionality allows us to do exactly that. I will probably add some more detail in the wiki on how to use it.