Remove blank line added during import of VBE Form
For some odd reason, importing a Forms2.0 object adds a blank line at the top of the code module with each import. This causes the module to slowly grow a blank section at the top. This update checks for and removes a single blank line at the top of the code module if it is found.