Add required parameter to DebugMode function to clarify dual purpose
The `DebugMode` function returns the current setting for the .BreakOnError option. What is not inherently intuitive is that it is *also* checking for any unhandled errors and either logging or stopping the code for developer review. As @A9G-Data-Droid pointed out, this is an undesirable affect that may lead to developers changing the code because they don't understand the purpose of a statement like "If DebugMode Then On Error Resume Next Else On Error Resume Next". By adding a required parameter, the parameter hint gives a clear indication of the secondary purpose of this function. Hopefully this will make the code more intuitive without adding cumbersome verbosity to the code base.