Use system codepage when reading export files from older db versions
When exporting source from a database in Microsoft Access 2000 format (version 4.0), objects are saved using the system codepage, and must be read as such when opening the file to sanitize it. Newer database versions export as UTF-8 automatically, so specifying the codepage is not necessary when reading the file. Fixes #217