Use original file format on build
When building from source, the add-in now attempts to use the same file format from the original database. This is important if you are maintaining databases in legacy versions. Also added error handling to give the user some additional clues on how to change the collation order if the database file creation fails. See #217