Generate initial translation template file
This file was automatically generated based on the strings that were automatically generated based on the captions of the form controls. There are still some minor adjustments that will need to be made, but this gives us a great start on a translation template file that can be edited in Poedit! :smile:  #87