Implement export conflict detection
This is a fairly major rework of the system, but allows us to confirm with the end user before overwriting changed files in the export folder. This is especially important in a multi-user environment where changes from another developer may appear in the source folder and could easily be overwritten by an export from the development database. This is also the foundation for being able to do a merge build where source changes are merged into an existing database without having to do a full build each time.