Integrate the clearing of orphaned files with the index
This allows us to confirm with the user before deleting files that don't have a corresponding database object. This could happen when a source file is added from another user but does not yet exist in the database. Without this change, an export would simply clear away the new file and it would never get into the working database.