Generate hash on UTF-8 string
The current system encoding must be taken into account when generating a hash on a text string. To make the hash more consistent with what you would expect from other programs or online generators, we are converting the text from the current system encoding to a UTF-8 byte array before passing it to the crypto APIs. Fixes #278