Remove any alternate source file format for TableDef export
Implementing  c4b0088 on dev branch.
If you were to change a local table to a linked table, the source file for the local table would still exist, which would then lead to an error during a build since a source file exists for both the local and the linked table. This update removes any existing alternate format during the export process. Fixes #296