Return hash from sanitize functions
Since we are already building the content and saving to a file, this is a great place to capture a hash of the content while it is still in memory. While this isn't ideal from an OOP perspective, it saves us from the performance hit of having to read the file from disk to compute the hash.