Refactor timer callback
We will likely be utilizing the timer for other functions going forward, so I consolidated the different timers into a single timer and callback function and added an enum to manage the different types of callbacks. This also moves the related code into close proximity which simplifies debugging and maintenance going forward.