Register COM dll
To properly install the COM dll, we need to register the class ID in the registry by calling regsvr32. This registers the class in the current user hive (I believe without requiring admin privileges) and calls the embedded code in the DLL to add the Access add-in registration so it shows up as an available COM add-in.