Rework execution handoff from local to add-in
When you use Application.run, it executes in the current database first, before an add-in database. This section allows us to move the execution over to the add-in with a clean handoff using the API Timer so we don't try to manipulate the local database from the add-in while code is still active in the local database. This mostly applies when building the add-in project itself, but is helpful for other areas as well.