Add feature to change repaint interval
Default repaint (screen updates) for the progress bar class is every second. But some projects (like this add-in) actually need more frequent updates to improve the user experience. This allows us to set the desired update interval after initializing the progress bar if we need something other than the default.