Update progress bar every half second
This restores the desired functionality that we previously had with the console output. Even though the progress bar was updating every second, the console was triggering updates every half second, giving better visual feedback on the progress of export operations. We consolidated some of the functions with the progress bar class rewrite, which did away with some redundant code, moving the update logic to the progress bar. This change brings back the half-second updates through the progress bar increments.