Scope error handling in FSO to object creation
If we check for unhandled errors every time we use the FSO function, we may throw an unhandled error message even though the calling code was intentionally trapping an error. As a case in point, if you tried to install the add-in while it was running in another version of Access, it would go to the debug window, rather than trapping the error and displaying the more intuitive message it was designed to show. We technically only need this error handling in FSO when creating the object, just in case we are unable to create it for some reason.