Add fallback to system meter for clsLblProg
If you use the label progress bar class without setting the labels through the Initialize method, it will now fall back to using the SysCmd progress meter in the application window instead. This improves the testability of code by allowing you to retain regular progress bar usage in code, but simply pass a progress bar class object without binding it to labels first. (I am using this functionality in other projects outside this add-in that benefit from this flexibility.)