Avoid exporting twice when scanning for changes
Some database objects have to be exported (and possibly sanitized) to determine if there are changes from the existing source files. But we don't know if we will actually use the temp file till after the user makes a decision on how to handle conflicts. This change allows us to stage the files used during the scan, and use them for the export if they are available before proceeding with a fresh export. For certain projects (such as large ADP projects) this can significantly reduce the export time.