Import themes before properties
The "Theme Resource Name" property needs to be set AFTER importing the theme files. Otherwise it can result in this property disappearing and the default theme being created as a duplicate the next time the database is opened. (Note that the theme is not available for use until after the database is reopened.)