Add new function to move source file(s)
Some component types export multiple source files. When exporting to a temp folder to check for changes, it is important that we move all the related source files back to the original folder if the user approves the conflict resolution. This refactor moves the logic to each component module, since that is where the decision is made on which types of files to export.