Avoid initializing main form when building add-in
Initializing forms ensures that the theme settings are properly applied to form objects when building from source. However, it is typically not necessary to do this on the main form of the add-in. (We could if we really wanted to, but it would require staging the add-in form while the CurrentDb version was opened in design view.)