Mark functions as Work In Progress
I am leaving the queue framework in place, since we may use that for other things in the future, but marking a couple functions as Work In Progress (WIP) since they don't currently work as intended when used with the queue. (Waiting for the queue persists the CodeProject as the current VBProject, preventing the worker from setting the current database as the current VBProject long enough for the DoCmd.DeleteObject to operate in the correct context. #383