Prevent merge of add-in forms
After significant testing and trying various approaches to replace a form in the current database that has the same name as a form in this add-in, I have come to the conclusion that it is not worth trying to support that functionality. This add-in can be built from source in a few seconds, so we really don't need to be able to merge form objects in this project. (All other database objects can be merged.) #383