Start building out git commands to split files
Looking at using this as a way to split form source code into two files, one with the VBA code, and the other with the form definition. The git commands would give us a way to upgrade an existing database to use the split file approach without losing the file history. #378