Refactor GetAllFromDB to return component class
This function typically returns a collection of the IDbComponent classes representing the actual components in the database. In the case of a component type with a single source file, only one of them is processed during the export/import operation. (That's why we set all of them to the current instance of the class.) However, the count of the dictionary returned by this function is used to determined the number of changed items in the database, so we should return an empty dictionary if there were no changed connection strings. (Otherwise it would always export the db-connection.json file, even when nothing has changed.) #388