Improve performance of testing for linked table
Checking the .Connect property is a quick and reliable way to determine if a table is a linked table, but it becomes a performance bottleneck when run 423 times scanning a large database with no changes. By caching the value and looking up the type from MSysObjects, I was able to cut the scanning time by nearly 50% on a large, complex database! Instead of spending over three seconds to check the tables, it is down to under a second for that object type, which remains the slowest object type to scan for changes in that database.