Dynamically scale columns on resize
Now that the conflict resolution form is now resizable, we can automatically scale the column widths to proportionately increase or decrease based on the size of the form. This way you can make the form wider to see more of the file names. Additionally, you can specify the names of some controls that you don't want to size because expanding them provides no value to the user since all the content is already displayed.