Change label hyperlinks to buttons
Label hyperlinks are a very handy and intuitive navigation aid, but they can cause problems when a database has a "Hyperlink Base" specified. Every time you click the navigation link, it opens a web page to the hyperlink base, even if you don't have any specific path defined. (This happens with text box links as well, but not in datasheet mode.) Changing these links to transparent command buttons with the CursorOnHover property allows us to simulate the same functionality without the actual hyperlink invocation when the hyperlink base is defined. (As it is on the testing database.)