Fix GetStringHash(with BOM) to match WriteFile
We made a change a while ago to always include a vbCrLf character at the end of each source file. This was done through the WriteFile function. But the GetStringHash did not use the same method when creating a hash to simulate a file, so the two didn't match. I updated the GetStringHash function to include the same logic as the WriteFile function so they will produce the same output. I also added a few tests to ensure that these are working as expected moving forward.