Adjust VBE form export to minimize noise
Described more fully in the comments, but we should only actually export the binary *.frx file when the VBE form design actually changes. (Otherwise we will see that file as different in every single full export.) Verified the new behavior in the testing database and it is working exactly as intended. :+1: