Add form to test extended character export
This form includes UTF-8 characters in the form controls, and extended characters in both the form and the underlying VBA code. This cannot be exported and imported on UTF-8 systems without splitting the VBA code from the form design. After finishing the functionality to split forms into two source files, this will serve as a test form to ensure that we can successfully build this type of form from source. (This happens more commonly in non-English environments, and is something that I would like to support.) #378