Rename function for COM add-in filename
This function had shared the same name as the public function in the installer. While this worked because of the overloading functionality where the private function defined in this module took precedence over the global function, it could really cause confusion down the road! I renamed the function to explicitly clarify which add-in we are talking about.  :smile: