Add worker to handle removing add-in files
The worker class gives us the ability to remove the actual database add-in file and related lock file during the uninstall process. The worker script is then able to delete itself, allowing us a clean uninstall of the add-in files. (The worker creates a vbscript file that runs independently of the Access VBA code, but can interact with the add-in for some operations. We don't use it a lot at this point, but it may be used more in the future.)