Merge dependent components after merge
Sometimes merging an object of one component type can affect another component type. For example, merging (deleting and importing) a table will also remove the custom description stored as a document property. The document property was unchanged at the beginning of the operation, but becomes changed by the time the merge is complete. This particular change merges any changes to the document component after merging any type of object visible in the navigation pane. (I discovered this effect when testing the merging using the Testing database.)