Fix logic error in relations export
Do not reset the persisted reference to the CurrentDB when retrieving the relation objects from the database. The "modified only" and "all objects" cached items both share the same parent reference to the current DB. If we reassign the CurrentDB reference, it will invalidate the object references of the previously assigned relation objects.