Drop unnecessary library prefix for Dictionary
There is no conflicting class in this project that would require us to explicitly specify the library for the Dictionary object. My general style preference is in keeping the code concise and not including library prefixes unless they are needed to prevent ambiguity. (Such as ADODB/DAO which both have a `recordset` object.)