Refactor table data listing/filtering in options
Refactored the code that pulls a list of the tables in the current database so the user can specify the tables from which they want to export data as part of the export process. Bitwise operators are not supported in Microsoft Access SQL queries, so this logic has been moved to VBA instead. Additionally, some system tables that cannot be viewed or modified have been excluded from the table listing. I also adjusted the number captions to reflect the number of items hidden/displayed in the current context. (This seemed to be the most logically intuitive number to display.) #403