Fix error logging within CatchAny function
The cascading call stack involved in logging an error also includes checking for unhandled errors, which was triggering a second error handling process before the first one completes. This change provides a state variable to monitor whether we are in the process of handling an error, and will not recursively trigger the error handling routines. #409