Use Unicode characters for three dots in ribbon
Regular period characters get stripped out if you have three in a row in a ribbon caption. I am using Unicode characters instead so I can still show "Build As..." in the caption. (If someone has a better way to do this, please let me know! I don't like this kind of clever hack, but I don't know of a better way to do it.)