Provide alternate fallback for sp_GetDDL
The ideal is to install sp_GetDDL at the server level, and run it as a system stored procedure against any database on the server. But this may not be possible in some environments. If you can't use it as a system SP, then it can be used as a local SP in the target database. If you can't install it at all, then you can fall back to built-in SQL functions that will produce some usable output. #415