Add option to SplitLayoutFromVBA
This option (on by default) will save the VBA code from forms and reports as a related .cls file. (Still under development.) #378

Also removed the "Strip out Publish Option" from the options form. I have never heard of a case where this needs to be changed, and it frees up space for the new option we are adding without cluttering the form.